src/backend/requirements.txt
Flask>=2.2,<3
pillow
pyautogui
requests
openai
python-dotenv
